Fix tests, and correct code to run these tests
