+++ acalib/core/__init__.py
@@ -1,5 +1,7 @@
 from .image_analysis import *
+from .spectra_analysis import *
 from .function_models import *
 from .utils import *
 from .transform import *
 from .compact_rep import *
+from .statistics import *
